src/main/java/com/baidu/oped/aqueducts/bolt/BoltB.java
/*
 * Copyright (C) 2014 Baidu, Inc. All Rights Reserved.
 */
package com.baidu.oped.aqueducts.bolt;

import backtype.storm.topology.BasicOutputCollector;
import backtype.storm.topology.OutputFieldsDeclarer;
import backtype.storm.topology.base.BaseBasicBolt;
import backtype.storm.tuple.Fields;
import backtype.storm.tuple.Tuple;
import backtype.storm.tuple.Values;

/**
 * Author: qudongfang@baidu.com
 * Date: 2014:10:30 11:01:00
 */

public class BoltB extends BaseBasicBolt {

    @Override
    public void execute(Tuple tuple, BasicOutputCollector basicOutputCollector) {
        // do sth
        String value = tuple.getStringByField("value");
        if (null == value || 0 == value.length()) {
            return;
        }

        value += ", processed by " + this.getClass().getName();

        basicOutputCollector.emit(new Values(value));
    }

    @Override
    public void declareOutputFields(OutputFieldsDeclarer outputFieldsDeclarer) {
        outputFieldsDeclarer.declare(new Fields("value"));
    }
}
